+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/plugins/abstractions/Chooser.kt
@@ -1,0 +1,10 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions
+
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
+
+interface Chooser : Plugin {
+    /**
+     * Must return postIds for posting
+     */
+    fun triggerChoose(): Collection<Int>
+}