+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/choosers/Chooser.kt
@@ -1,11 +1,12 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers
 
-import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser
 
-
-interface Chooser : Plugin {
-    /**
-     * Must return postIds for posting
-     */
-    fun triggerChoose(): Collection<Int>
-}
+@Deprecated(
+    "Must be replaced",
+    ReplaceWith(
+        "Chooser",
+        "com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser"
+    )
+)
+typealias Chooser = Chooser